+++ pkg/rules/benchmark/benchmark_test.go
@@ -6,10 +6,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/aquasecurity/tracee/tracee-rules/benchmark/signature/golang"
-	"github.com/aquasecurity/tracee/tracee-rules/benchmark/signature/rego"
-	"github.com/aquasecurity/tracee/tracee-rules/benchmark/signature/wasm"
-	"github.com/aquasecurity/tracee/tracee-rules/engine"
+	"github.com/aquasecurity/tracee/pkg/rules/benchmark/signature/golang"
+	"github.com/aquasecurity/tracee/pkg/rules/benchmark/signature/rego"
+	"github.com/aquasecurity/tracee/pkg/rules/benchmark/signature/wasm"
+	"github.com/aquasecurity/tracee/pkg/rules/engine"
 	"github.com/aquasecurity/tracee/types"
 	"github.com/stretchr/testify/require"
 )
