+++ pkg/rules/benchmark/helpers_test.go
@@ -10,7 +10,7 @@ import (
 	"os"
 
 	"github.com/aquasecurity/tracee/pkg/external"
-	"github.com/aquasecurity/tracee/tracee-rules/engine"
+	"github.com/aquasecurity/tracee/pkg/rules/engine"
 	"github.com/aquasecurity/tracee/types"
 )
