+++ pkg/rules/benchmark/signature/golang/anti_debugging.go
