+++ pkg/rules/benchmark/signature/golang/code_injection.go
