+++ pkg/rules/benchmark/signature/golang/noop.go
