+++ pkg/rules/benchmark/signature/rego/anti_debugging_ptraceme.rego
