+++ pkg/rules/benchmark/signature/rego/code_injection.rego
