+++ pkg/rules/benchmark/signature/rego/signatures.go
@@ -3,7 +3,7 @@ package rego
 import (
 	_ "embed"
 
-	"github.com/aquasecurity/tracee/tracee-rules/regosig"
+	"github.com/aquasecurity/tracee/pkg/rules/regosig"
 	"github.com/aquasecurity/tracee/types"
 	"github.com/open-policy-agent/opa/compile"
 )
