+++ pkg/rules/benchmark/signature/wasm/anti_debugging_ptraceme.rego
