+++ pkg/rules/benchmark/signature/wasm/code_injection.rego
