+++ pkg/rules/benchmark/signature/wasm/signatures.go
