+++ pkg/rules/engine/engine.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 
 	"github.com/aquasecurity/tracee/pkg/external"
-	"github.com/aquasecurity/tracee/tracee-rules/metrics"
+	"github.com/aquasecurity/tracee/pkg/rules/metrics"
 	"github.com/aquasecurity/tracee/types"
 	"github.com/open-policy-agent/opa/ast"
 )
