+++ pkg/rules/metrics/stats.go
