+++ pkg/rules/regosig/aio_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	"github.com/aquasecurity/tracee/pkg/external"
+	"github.com/aquasecurity/tracee/pkg/rules/regosig"
 	"github.com/aquasecurity/tracee/signatures/signaturestest"
-	"github.com/aquasecurity/tracee/tracee-rules/regosig"
 	"github.com/aquasecurity/tracee/types"
 	"github.com/open-policy-agent/opa/compile"
 	"github.com/stretchr/testify/assert"
