+++ pkg/rules/regosig/common_test.go
