+++ pkg/rules/regosig/mapper.go
