+++ pkg/rules/regosig/mapper_test.go
